Refactor registry to make multihead support reusable
